Use new app for ERC4337 related logic
