+++ safe_transaction_service/account_abstraction/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AccountAbstractionConfig(AppConfig):
+    name = "safe_transaction_service.account_abstraction"
+    verbose_name = "Account Abstraction (ERC4337) support for Safe Transaction Service"
