+++ safe_transaction_service/account_abstraction/tests/mocks/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa F401
+from .token_list_mock import token_list_mock
