+++ safe_transaction_service/account_abstraction/tests/test_models.py
@@ -1,0 +1,5 @@
+from django.test import TestCase
+
+
+class TestModels(TestCase):
+    pass
