+++ safe_transaction_service/account_abstraction/urls.py
@@ -1,0 +1,3 @@
+app_name = "account_abstraction"
+
+urlpatterns = []
